LayoutTests/imported/w3c/web-platform-tests/workers/shared-worker-name-via-options-expected.txt


FAIL Test the name property of shared workers mixing constructor options and constructor strings Error: assert_equals: expected 2 but got 1
